Save Image with button.
Everything is alright. Here it has to be an alert that will notify the
user that the image is saved.
Tried but it was super hard. I will submit it though If anyone wants to
try out. Possible to make other functions like "Delete Image", "Report
Image" etc. That's why the icon is an "Options" icon.